refactor(test): rename `test` directory to `__tests__`
