+++ __tests__/data/svg.ts
