+++ __tests__/esm/attributes-to-props.test.mjs
